pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argparse>=1.4.0",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pytest-cov>=6.0.0",
    "pytest>=8.3.4",
    "ruff>=0.9.4",
]

[project.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]
